Create Task -1
This Python program functions as a basic calculator, allowing users to perform simple arithmetic operations. Upon running the program, users are prompted to select an operator from a list of available options, such as addition, subtraction, multiplication, and division. Once the operator is selected, the program requests the user to input two numbers. It then performs the chosen operation on the provided numbers and displays the result. This straightforward calculator is ideal for quick and simple calculations.